fix: default를 인기순으로 바꿔 sortOption 함수 리액트 쿼리함수 내부로 이동시키고 default값을 POPULAR로 수정
